+++ core/ProfileDetailGenerator.py
@@ -1,25 +1,42 @@
 import logging
 import re
-from collections.abc import Mapping
+from collections.abc import Callable
+from typing import Mapping, List, TypedDict, Any, Optional
+
+from core.exceptions.profile import MissingProfileException
+from model.UiDataModel import TranslationElementDisplay
+from model.dse import FieldDetail, ProfileDetail, Filter
+from util.fhir.enums import FhirPrimitiveDataType
+
+Profile = Mapping[str, any]
 
 
-class ProfileDetailGenerator():
+class SearchParamPathMapping(TypedDict):
+    search_param: str
+    fhir_path: str
 
-    def __init__(self, profiles, mapping_type_code, blacklistedValueSets, fields_to_exclude, field_trees_to_exclude, reference_base_url):
 
-        self.logger = logging.getLogger(self.__class__.__name__)
-        self.blacklistedValueSets = blacklistedValueSets
+class ProfileDetailGenerator:
+    __logger = logging.getLogger(__name__)
+    blacklisted_values_sets: List[str]
+    profiles: Mapping[str, Mapping[str, Mapping[str, any]]]
+    mapping_type_code: Mapping[str, SearchParamPathMapping]
+    fields_to_exclude: List[str]
+    fields_trees_to_exclude: List[str]
+    reference_base_url: str
+
+    def __init__(self, profiles, mapping_type_code, blacklisted_value_sets, fields_to_exclude, field_trees_to_exclude,
+                 reference_base_url):
+        self.blacklisted_value_sets = blacklisted_value_sets
         self.profiles = profiles
+        # Prevents having to generate the mapping over and over again
+        self.__all_profiles = self.__get_profiles()
         self.mapping_type_code = mapping_type_code
         self.fields_to_exclude = fields_to_exclude
         self.field_trees_to_exclude = field_trees_to_exclude
-        self.simple_data_types = ["instant", "time", "date", "dateTime", "decimal", "boolean", "integer", "string",
-                                  "uri", "base64Binary", "code", "id", "oid", "unsignedInt", "positiveInt", "markdown",
-                                  "url", "canonical", "uuid"]
         self.reference_base_url = reference_base_url
 
-    def find_and_load_scruct_def_from_path(self, struct_def, path):
-
+    def find_and_load_struct_def_from_path(self, struct_def: Mapping[str, any], path: str):
         elements = struct_def['snapshot']["element"]
 
         for elem in (elem for elem in elements if elem['id'] == path):
@@ -35,12 +52,12 @@ class ProfileDetailGenerator():
                 if not target_profile_url:
                     continue
 
-                if target_profile_url not in self.profiles:
-                    for profile in self.profiles:
+                if target_profile_url not in self.__all_profiles:
+                    for profile in self.__all_profiles:
                         if target_profile_url.split("/")[-1] == profile.split("/")[-1]:
-                            return self.profiles[profile]['structureDefinition']
+                            return self.__all_profiles[profile]['structureDefinition']
 
-                return self.profiles[target_profile_url]['structureDefinition']
+                return self.__all_profiles[target_profile_url]['structureDefinition']
 
         return None
 
@@ -51,13 +68,14 @@ class ProfileDetailGenerator():
         part_match = re.search(r'\((.*?)\)', fhir_path)
 
         if part_match:
-            struct_def = self.find_and_load_scruct_def_from_path(struct_def, part_match.group(1))
+            struct_def = self.find_and_load_struct_def_from_path(struct_def, part_match.group(1))
 
             if not struct_def:
                 return None
 
-            type = struct_def['type']
-            return self.get_value_sets_for_code_filter(struct_def, f"{type}.{fhir_path.split(').')[-1]}")
+            struct_def_type = struct_def['type']
+            return self.get_value_sets_for_code_filter(struct_def,
+                                                       f"{struct_def_type}.{fhir_path.split(').')[-1]}")
 
         pattern = rf"{fhir_path}[^.]*$"
 
@@ -67,7 +85,7 @@ class ProfileDetailGenerator():
 
                 value_set = elem["binding"]["valueSet"]
 
-                if value_set not in self.blacklistedValueSets:
+                if value_set not in self.blacklisted_value_sets:
                     value_sets.append(value_set)
 
             elif len(value_sets) == 0 and "patternCoding" in elem or "fixed" in elem:
@@ -84,7 +102,7 @@ class ProfileDetailGenerator():
 
                 value_set = elem["binding"]["valueSet"]
 
-                if value_set not in self.blacklistedValueSets:
+                if value_set not in self.blacklisted_value_sets:
                     value_sets.append(value_set)
 
             elif "patternCoding" in elem or "fixed" in elem:
@@ -95,9 +113,10 @@ class ProfileDetailGenerator():
 
         return value_sets
 
-    def get_field_in_node(self, node, id_end):
+    @staticmethod
+    def get_field_in_node(node, id_end):
 
-        if "children" not in node:
+        if "children" not in node or node["children"] is None:
             node["children"] = []
 
         children = node["children"]
@@ -105,77 +124,94 @@ class ProfileDetailGenerator():
         for index in range(0, len(children)):
 
             child = children[index]
-            if child["id"].endswith(id_end):
+            if child.id.endswith(id_end):
                 return index
 
         return -1
 
-    def insert_field_to_tree(self, tree, field):
-
+    def insert_field_to_tree(self, tree: FieldDetail, field: FieldDetail):
         cur_node = tree
-        path = re.split(r"[.:]", field["id"])
+        path = re.split(r"[.:]", field.id)
         path = path[1:]
         parent_recommended = False
-        field_type = field.get('type')
+        field_type = field.type
 
         if len(path) == 0:
             return
 
-        if field_type in self.simple_data_types and len(path) > 1:
+        # TODO: This is a temporary workaround to allow both the postal code and the country information to be selected
+        #       during data selection. To preserve context, selecting elements with simple data types which are not on
+        #       the top level of a resource is disabled (e.g. to forbid selecting just Coding.code without
+        #       Coding.system etc.).
+        #       In the future we should switch to a more dynamic solution were the selectable elements can be defined in
+        #       externalized config files using a well-defined syntax to prevent such hard-coded solutions.
+        if (field_type in FhirPrimitiveDataType and len(path) > 1 and
+                field.id not in {"Patient.address:Strassenanschrift.postalCode",
+                                        "Patient.address:Strassenanschrift.country"}):
             return
 
         for index in range(0, len(path) - 1):
 
             id_end = path[index]
-            field_child_index = self.get_field_in_node(cur_node, id_end)
+            field_child_index = self.get_field_in_node(cur_node.__dict__, id_end)
 
             if field_child_index != -1:
-                cur_node = cur_node["children"][field_child_index]
+                cur_node = cur_node.children[field_child_index]
 
-            if parent_recommended == False and "recommended" in cur_node:
-                parent_recommended = cur_node['recommended']
+            if not parent_recommended:
+                parent_recommended = cur_node.recommended
 
             if parent_recommended:
-                field['recommended'] = False
+                field.recommended = False
 
-        if "children" not in cur_node:
-            cur_node["children"] = []
+        if cur_node.children is None:
+            cur_node.children = []
 
-        cur_node["children"].append(field)
+        cur_node.children.append(field)
 
-    def filter_element(self, element):
+    def filter_element(self, element: Mapping[str, any]) -> bool:
+        # TODO: This is a temporary workaround to allow both the postal code and the country information to be selected
+        #       during data selection. To preserve context, selecting elements with simple data types which are not on
+        #       the top level of a resource is disabled (e.g. to forbid selecting just Coding.code without
+        #       Coding.system etc.).
+        #       In the future we should switch to a more dynamic solution were the selectable elements can be defined in
+        #       externalized config files using a well-defined syntax to prevent such hard-coded solutions.
+        element_id = element.get('id')
+        if element_id in {"Patient.address:Strassenanschrift.postalCode", "Patient.address:Strassenanschrift.country"}:
+            return False
 
         attributes_true_level_one = ["mustSupport", "isModifier", "min"]
 
         if all(element.get(attr) is False or element.get(attr) == 0 or attr not in element
                for attr in attributes_true_level_one):
-            self.logger.debug(f"Excluding: {element['id']} as not mustSupport, modifier or min > 0")
+            self.__logger.debug(f"Excluding: {element['id']} as not mustSupport, modifier or min > 0")
             return True
 
         attributes_true_level_two = ["mustSupport", "isModifier"]
 
         if all(element.get(attr) is False or element.get(attr) == 0 or attr not in element
                for attr in attributes_true_level_two) and len(element["id"].split(".")) > 2:
-            self.logger.debug(f"Excluding: {element['id']} as not mustSupport or modifier on level > 2")
+            self.__logger.debug(f"Excluding: {element['id']} as not mustSupport or modifier on level > 2")
             return True
 
         if any(element['id'].endswith(field) or f"{field}." in element['id']for field in self.fields_to_exclude):
-            self.logger.debug(f"Excluding: {element['id']} as excluded field")
+            self.__logger.debug(f"Excluding: {element['id']} as excluded field")
             return True
 
         if any(f"{field}" in element['id'] for field in self.field_trees_to_exclude):
-            self.logger.debug(f"Excluding: {element['id']} as part of field tree")
+            self.__logger.debug(f"Excluding: {element['id']} as part of field tree")
             return True
 
         if "[x]" in element['id'] and not element['id'].endswith("[x]"):
-            self.logger.debug(f"Excluding: {element['id']} as sub-elements relevant")
+            self.__logger.debug(f"Excluding: {element['id']} as sub-elements relevant")
             return True
 
         if element["base"]["path"].split(".")[0] in {"Resource", "DomainResource"} and not "mustSupport" in element:
-            self.logger.debug(f"Excluding: {element['id']} as base is Resource or DomainResource and not must Support")
+            self.__logger.debug(f"Excluding: {element['id']} as base is Resource or DomainResource and not mustSupport")
             return True
 
-    def check_at_least_one_in_elem_and_true(self, element, attributes_to_check):
+    @staticmethod
+    def check_at_least_one_in_elem_and_true(element: Mapping[str, any], attributes_to_check: List[str]) -> bool:
 
         path = element["id"].split(".")
 
@@ -188,9 +224,10 @@ class ProfileDetailGenerator():
 
         return True
 
-    def get_name_from_id(self, id):
+    @staticmethod
+    def get_name_from_id(element_id: str) -> str:
 
-        name = id.split(".")[-1]
+        name = element_id.split(".")[-1]
         name = name.split(":")[-1]
         name = name.replace("[x]", "")
 
@@ -208,191 +245,372 @@ class ProfileDetailGenerator():
 
         return None
 
-    def get_element_by_content_ref(self, content_ref, elements):
-
+    @staticmethod
+    def get_element_by_content_ref(content_ref: str, elements: List[Mapping[str, any]]) -> Optional[Mapping[str, any]]:
         for element in elements:
             if element['id'] == content_ref:
                 return element
-
         return None
 
-    def get_value_for_lang_code(self, data, langCode):
+    @staticmethod
+    def get_value_for_lang_code(data: Mapping[str, any], lang_code: str) -> str:
         for ext in data.get('extension', []):
-            if any(e.get('url') == 'lang' and e.get('valueCode') == langCode for e in ext.get('extension', [])):
+            if any(e.get('url') == 'lang' and e.get('valueCode') == lang_code for e in ext.get('extension', [])):
                 return next(e['valueString'] for e in ext['extension'] if e.get('url') == 'content')
         return ""
 
-    def resource_type_to_date_param(self, resource_type):
-
+    @staticmethod
+    def resource_type_to_date_param(resource_type: str) -> str:
         if resource_type == 'Condition':
             return "recorded-date"
 
         return "date"
 
-    def find_mii_references_by_type(self, fhir_type):
+    def __get_profiles(self, scope: Optional[str] = None) -> Mapping[str, Mapping[str, Any]]:
+        """
+        Returns all profile entries in a certain scope or all if none is provided
+        :param scope: Scope from which to return the profile entries
+        :return: All profile entries matching the provided scope
+        """
+        if scope:
+            return self.profiles.get(scope, {})
+        else:
+            return {k: v for d in self.profiles.values() for k, v in d.items()}
+
+    def __find_mii_references_by_type(self, fhir_type):
 
         matching_mii_references = []
 
-        for profile in self.profiles.values():
+        for profile in self.__all_profiles.values():
 
             profile_fhir_type = profile["structureDefinition"]["type"]
             if profile_fhir_type == fhir_type:
-                matching_mii_references.append(profile["url"])
+                matching_mii_references.extend(self.__resolve_selectable_profiles(profile["url"]))
 
         return matching_mii_references
 
-    def get_referenced_mii_profiles(self, element, field_type):
+    def __resolve_selectable_profiles(self, profile_url: str, module: Optional[str]=None) -> List[str]:
+        """
+        Collects profile URLs of all selectable profiles given a parent profile from which other profiles in the scope
+        of this class instance (all profiles in the ProfileDetailGenerator::profiles field) might be derived. If they
+        are then they would also be selectable if the parent profile defines the range of a Reference element
+        :param profile_url: string representing the URL of the profile from which to start the resolution
+        :param module: Optional module name string from which the profile identified by the `profile_url` parameter
+                       originates
+        :return: list of URLs of selectable profiles
+        """
+        selectable_profiles = []
+        children = filter(lambda t: t[1].get('baseDefinition', None) == profile_url,
+                          map(lambda p: (p.get('module'), p.get('structureDefinition', {})),
+                              self.__all_profiles.values()))
 
+        # FIXME: The logic should only rely on the value of the `abstract` element of a given StructureDefinition
+        #        instance to determine whether it itself is selectable or not. Currently whether other profiles in the
+        #        same module are derived from it is used to determine "abstractness"
+        # Determine whether the parent profile itself should be included
+        if profile_url in self.__all_profiles:
+            parent_profile = self.__all_profiles.get(profile_url, {}).get('structureDefinition', None)
+            if parent_profile:
+                if not parent_profile.get('abstract', False):
+                    if all(m != module for m, _ in children):
+                        selectable_profiles.append(profile_url)
+        elif profile_url not in self.__all_profiles:
+            # TODO: Decide on right log level since this matches every time the FHIR base resource profiles are
+            #       encountered
+            # If the URL does not match any profile in any scope it might be missing
+            self.__logger.debug(f"Provided profile URL '{profile_url}' is not present and cannot be analyzed further. "
+                                f"Consider including it via dependencies if this is not intended.")
+
+        for child_module, child in map(lambda p: (p.get('module'), p.get('structureDefinition', {})),
+                                       self.__all_profiles.values()):
+            if child.get('baseDefinition', None) == profile_url:
+                child_profile_url = child.get('url', None)
+                child_selectable_profiles = self.__resolve_selectable_profiles(child_profile_url, child_module)
+                selectable_profiles.extend(child_selectable_profiles)
+        return selectable_profiles
+
+    def get_referenced_mii_profiles(self, element: Mapping[str, any], field_type: str,
+                                    is_root: bool = True) -> List[str]:
+        """
+        Searches for and aggregates referenced MII profiles listed by Reference elements supporting and present within
+        the scope of profiles known to this ProfileDetailGenerator instance. Given some target profile specified for a
+        Reference element all MII profiles are identified that are its descendants and have no profiles based on them
+        in the instances profile scope
+        :param element: ElementDefinition instance for which to find MII profiles it supports
+        :param field_type: Shall be either 'Reference' if it supports the Reference FHIR data type or 'Extension' if it
+                           supports 'Extension' elements that support the Reference FHIR data type in their 'value'
+                           element
+        :param is_root: If set to 'True' the call is assumed to be the initial call of this function. It is used
+                        internally and should not be set by externally
+        :return: List of MII profile URLs supported by the element
+        """
         mii_references = []
+        supports_reference = False
 
-        if field_type == 'Reference':
+        match field_type:
+            case 'Reference':
+                supports_reference = True
+                target_profiles = []
 
-            target_profiles = None
+                for element_type in element['type']:
+                    target_profiles = element_type.get('targetProfile', [])
 
-            for type in element['type']:
-                if "targetProfile" in type:
-                    target_profiles = type['targetProfile']
-
-            for profile in target_profiles:
-                if profile.startswith('https://www.medizininformatik-initiative.de'):
-                    mii_references.append(profile)
-
-            if len(mii_references) == 0:
                 for profile in target_profiles:
-                    fhir_type = profile.rstrip('/').split('/')[-1]
-                    mii_references.extend(self.find_mii_references_by_type(fhir_type))
+                    mii_references.extend(filter(
+                        lambda p: p.startswith('https://www.medizininformatik-initiative.de'),
+                        self.__resolve_selectable_profiles(profile)
+                    ))
+
+                #if len(mii_references) == 0:
+                #    for profile in target_profiles:
+                #        fhir_type = profile.rstrip('/').split('/')[-1]
+                #        mii_references.extend(self.find_mii_references_by_type(fhir_type))
+
+            case 'Extension':
+                # Iterate over all types with type code 'Extension'
+                for ext_type in filter(lambda t: t.get('code', None) == "Extension", element.get('type', [])):
+                    # Iterate over all MII target profiles explicitly supported by the element
+                    for ext_profile_url in ext_type.get("profile", []):
+                        ext_profile = self.__all_profiles.get(ext_profile_url, None)
+                        if not ext_profile:
+                            raise MissingProfileException(f"Missing extension profile with URL '{ext_profile_url}' in "
+                                                          f"current scope. This can likely be fixed by adding its "
+                                                          f"snapshot to the DSE package snapshots directory")
+                        ext_elements = ext_profile.get('structureDefinition', {}).get('snapshot', {}).get('element', [])
+                        ext_value_elements = list(
+                            filter(lambda e: e.get('path', "").startswith("Extension.value"), ext_elements)
+                        )
+                        # By definition Extension instances can only have up to one 'value' element
+                        ext_value_element = ext_value_elements[0] if len(ext_value_elements) > 0 else None
+                        if ext_value_element and ext_value_element.get('max') != "0":
+                            # Iterate over all types supported by the extensions value element to retrieve MII profiles
+                            # they support
+                            for element_type in [t.get('code') for t in ext_value_element.get('type', [])]:
+                                references = self.get_referenced_mii_profiles(ext_value_element, element_type, False)
+                                mii_references.extend(references)
+
+                        # Filter for Extensions element definition in the Extensions definition
+                        ext_ext_elements = list(
+                            filter((lambda e: e.get('path', "").startswith('Extension.extension')),
+                                   ext_elements)
+                        )
+                        if len(ext_value_elements) > 0:
+                            # There are two cases we have to cover:
+                            # 1) There are references to other Extension profiles within some extension element
+                            #    definitions
+                            # 2) The Extension element is defined directly in the current Extension profile
+                            for ext_ext_element in ext_ext_elements:
+                                ext_ext_element_types = {t.get('code') for t in ext_ext_element.get('type', [])}
+                                references = []
+                                if 'Extension' in ext_ext_element_types:
+                                    references.extend(self.get_referenced_mii_profiles(ext_ext_element, "Extension",
+                                                                                       False))
+                                elif 'Reference' in ext_ext_element_types:
+                                    references.extend(self.get_referenced_mii_profiles(ext_ext_element, "Reference",
+                                                                                       False))
+                                mii_references.extend(references)
+
+        if is_root and supports_reference and len(mii_references) == 0:
+            self.__logger.warning(f"{field_type} element '{element.get('id')}' does not support any selectable MII "
+                                  f"profile in the current scope. This could be fixed by adding snapshots to the "
+                                  f"DSE package snapshots directory, but can also be an indicator that no MII profile "
+                                  f"is in the reference chain at all")
+            return []
 
         return mii_references
 
-    def generate_detail_for_profile(self, profile):
-        self.logger.info(f"Generating profile detail for: {profile['url']}")
+    def generate_detail_for_profile(self, profile: Mapping[str, any],
+                                    profile_tree: Optional[Mapping[str, any]]=None) -> Optional[ProfileDetail]:
+        profile_url = profile.get('url')
+        self.__logger.info(f"Generating profile detail [url='{profile_url}']")
 
-        if not "snapshot" in profile["structureDefinition"]:
-            self.logger.warning(f'profile with url {profile["url"]} has no snapshot - ignoring')
-            return None
+        try:
+            if not "snapshot" in profile["structureDefinition"]:
+                self.__logger.warning(f"Profile has no snapshot [url='{profile_url}'] => Skipping")
+                return None
 
-        struct_def = profile["structureDefinition"]
+            struct_def = profile["structureDefinition"]
+            date_param = self.resource_type_to_date_param(struct_def['type'])
 
-        date_param = self.resource_type_to_date_param(struct_def['type'])
+            if profile_tree and not self.__is_profile_selectable(profile_url, profile_tree):
+                self.__logger.debug(f"Profile is not selectable according to profile tree [url='{profile_url}'] => "
+                                    f"Skipping")
+                return None
 
-        profile_detail = {
-            "url": profile["url"],
-            "display": {"original": struct_def.get("title", ""),
-                        "translations": [
+            profile_detail = ProfileDetail(
+                url=profile_url,
+                display=TranslationElementDisplay(
+                    original=struct_def.get("title", ""),
+                    translations=[
+                        {
+                            "language": "de-DE",
+                            "value": self.get_value_for_lang_code(struct_def.get("_title", {}), "de-DE")
+                        },
+                        {
+                            "language": "en-US",
+                            "value": self.get_value_for_lang_code(struct_def.get("_title", {}), "en-US")
+                        }
+                    ]
+                ),
+                filters=[
+                    Filter(type="date", name=date_param, ui_type="timeRestriction")
+                ]
+            )
+
+            profile_type = profile['structureDefinition']['type']
+            code_search_param = (result := self.mapping_type_code.get(profile_type, None)) and result.get("search_param",
+                                                                                                          None)
+            fhir_path = (result := self.mapping_type_code.get(profile_type, None)) and result.get("fhir_path", None)
+            value_set_urls = None
+
+            if fhir_path is not None:
+                value_set_urls = self.get_value_sets_for_code_filter(profile['structureDefinition'], fhir_path)
+
+            if value_set_urls:
+                profile_detail.filters.append(Filter(
+                    type="token",
+                    name=code_search_param,
+                    ui_type="code",
+                    valueSetUrls=value_set_urls,
+                ))
+
+            field_tree = FieldDetail(id=struct_def.get('type'))
+            source_elements = profile["structureDefinition"]["snapshot"]["element"]
+
+            for element in source_elements:
+                field_id = element["id"]
+
+                if self.filter_element(element):
+                    continue
+
+                if 'contentReference' in element:
+                    content_reference_split = element['contentReference'].split("#")
+                    if len(content_reference_split) > 1:
+                        content_reference = content_reference_split[1]
+                    else:
+                        content_reference = content_reference_split[0]
+
+                    element = self.get_element_by_content_ref(content_reference, source_elements)
+
+                field_type = None
+
+                if "type" in element:
+                    for element_type in element["type"]:
+
+                        field_type = element_type["code"]
+
+                        if element_type["code"] == "Reference":
+                            break
+
+                else:
+                    self.__logger.warning(f"Element '{element.get('id')}' without type => Discarding")
+                    continue
+
+                is_recommended_field = self.check_at_least_one_in_elem_and_true(element, ["min"])
+                is_required_field = self.check_at_least_one_in_elem_and_true(element, ["isModifier"])
+
+                # FIXME: Temporary fix to make elements of type 'Reference' not recommended due to issues in the UI
+                #        except for references to the MII Medication profile
+                if field_type == "Reference" and not ".medication" in field_id:
+                    is_recommended_field = False
+
+                name = self.get_name_from_id(element["id"])
+
+                try:
+                    referenced_mii_profiles = self.get_referenced_mii_profiles(element, field_type)
+                except MissingProfileException as exc:
+                    self.__logger.warning(f"Could not resolve referenced MII profiles [element_id='{field_id}']. "
+                                          f"Reason: {exc}")
+                    referenced_mii_profiles = []
+
+                field = FieldDetail(
+                    id=field_id,
+                    display=TranslationElementDisplay(
+                        original=name,
+                        translations=[
+                             {
+                                 "language": "de-DE",
+                                 "value": self.get_value_for_lang_code(element.get('_short', {}), "de-DE")
+                             },
+                             {
+                                 "language": "en-US",
+                                 "value": self.get_value_for_lang_code(element.get('_short', {}), "en-US")
+                             }
+                        ],
+                    ),
+                    description=TranslationElementDisplay(
+                        original=element.get("definition", ""),
+                        translations=[
                             {
                                 "language": "de-DE",
-                                "value": self.get_value_for_lang_code(struct_def.get("_title", {}), "de-DE")
+                                "value": self.get_value_for_lang_code(element.get('_definition', {}), "de-DE")
                             },
                             {
                                 "language": "en-US",
-                                "value": self.get_value_for_lang_code(struct_def.get("_title", {}), "en-US")
+                                "value": self.get_value_for_lang_code(element.get('_definition', {}), "en-US")
                             }
                         ]
-                        },
-            "filters": [
-                {"type": "date", "name": date_param, "ui_type": "timeRestriction"}
-            ],
-        }
+                    ),
+                    referencedProfiles=referenced_mii_profiles,
+                    type=field_type,
+                    recommended=is_recommended_field,
+                    required=is_required_field
+                )
 
-        profile_type = profile['structureDefinition']['type']
-        code_search_param = (result := self.mapping_type_code.get(profile_type, None)) and result.get("search_param",
-                                                                                                      None)
-        fhir_path = (result := self.mapping_type_code.get(profile_type, None)) and result.get("fhir_path", None)
-        value_set_urls = None
+                if field_type == "Reference" and len(referenced_mii_profiles) == 0:
+                    self.__logger.warning(f"Element '{element.get('id')}' references profiles that do not match any "
+                                          f"MII profile => Discarding")
+                    continue
 
-        if fhir_path is not None:
-            value_set_urls = self.get_value_sets_for_code_filter(profile['structureDefinition'], fhir_path)
+                self.insert_field_to_tree(field_tree, field)
 
-        if value_set_urls:
-            profile_detail['filters'].append({
-                "type": "token",
-                "name": code_search_param,
-                "ui_type": "code",
-                "valueSetUrls": value_set_urls,
-            })
+            profile_detail.fields = field_tree.children
 
-        field_tree = {"children": []}
-        source_elements = profile["structureDefinition"]["snapshot"]["element"]
+            return profile_detail
+        except Exception as exc:
+            raise Exception(f"Failed to generate profile details for profile '{profile.get('url')}'") from exc
 
-        for element in source_elements:
+    @staticmethod
+    def __is_profile_selectable(profile_url: str, profile_tree: Mapping[str, any]) -> bool:
+        """
+        Searches the profile tree to determine whether a profile (identified by the provided URL) is selectable. Returns
+        False if the profile is not present in the tree or has no attribute 'selectable'
+        :param profile_url: URL of the profile for which to determine whether it is selectable
+        :param profile_tree: Profile tree to search
+        :return: Boolean indicating whether profile is selectable
+        """
+        if profile_tree.get('url') == profile_url:
+            return profile_tree.get('selectable', False)
+        else:
+            return any([ProfileDetailGenerator.__is_profile_selectable(profile_url, tree)
+                        for tree in profile_tree.get('children', [])])
 
-            field_id = element["id"]
-
-            if self.filter_element(element):
-                continue
-
-            if 'contentReference' in element:
-                content_reference_split = element['contentReference'].split("#")
-                if len(content_reference_split) > 1:
-                    content_reference = content_reference_split[1]
-                else:
-                    content_reference = content_reference_split[0]
-
-                element = self.get_element_by_content_ref(content_reference, source_elements)
-
-            field_type = None
-
-            if "type" in element:
-                for type in element["type"]:
-
-                    field_type = type["code"]
-
-                    if type["code"] == "Reference":
-                        break
-
+    def generate_profile_details_for_profiles_in_scope(self, scope: str,
+                                                       cond: Optional[Callable[[Mapping[str, Any]], bool]] = None,
+                                                       profile_tree: Mapping[str, any] = None
+                                                       ) -> List[ProfileDetail]:
+        """
+        Generate profile details for all profiles within the given scope
+        :param scope: The scope containing all the profiles to generated details for
+        :param cond: Optional filter to match only certain StructureDefinition instances in scope
+        :param profile_tree: Optional profile tree to determine whether a profile detail should be generated for a given
+                             profile based on whether it is selectable
+        :return: List of profile details for the given scope
+        """
+        self.__logger.info(f"Generating profile details for profiles in scope '{scope}'")
+        if cond is None:
+            def true(_: Any): return True
+            cond = true
+        profile_details = []
+        profiles_in_scope = self.__get_profiles(scope)
+        if len(profiles_in_scope) == 0:
+            self.__logger.warning(f"No profiles in scope '{scope}' => No profile details will be generated")
+            return profile_details
+        for profile_entry in profiles_in_scope.values():
+            sd = profile_entry.get("structureDefinition", {})
+            if cond(sd):
+                if profile_detail := self.generate_detail_for_profile(profile_entry, profile_tree):
+                    profile_details.append(profile_detail)
             else:
-                self.logger.warning(f"Element without type: {element} - discarding")
-                continue
-
-            is_recommended_field = self.check_at_least_one_in_elem_and_true(element, ["min"])
-            is_required_field = self.check_at_least_one_in_elem_and_true(element, ["isModifier"])
-
-            # FIXME: Temporary fix to make elements of type 'Reference' not recommended due to issues in the UI except
-            #        for references to the MII Medication profile
-            if field_type == "Reference" and not ".medication" in field_id:
-                is_recommended_field = False
-
-            name = self.get_name_from_id(element["id"])
-
-            field = {"id": field_id,
-                     "display": {"original": name,
-                                 "translations": [
-                                     {
-                                         "language": "de-DE",
-                                         "value": self.get_value_for_lang_code(element.get('_short', {}), "de-DE")
-                                     },
-                                     {
-                                         "language": "en-US",
-                                         "value": self.get_value_for_lang_code(element.get('_short', {}), "en-US")
-                                     }
-                                 ],
-                                 },
-                     "description": {"original": element.get("definition", ""),
-                                     "translations": [
-                                         {
-                                             "language": "de-DE",
-                                             "value": self.get_value_for_lang_code(element.get('_definition', {}),
-                                                                                   "de-DE")
-                                         },
-                                         {
-                                             "language": "en-US",
-                                             "value": self.get_value_for_lang_code(element.get('_definition', {}),
-                                                                                   "en-US")
-                                         }
-                                     ],
-                                     },
-                     "referencedProfiles": self.get_referenced_mii_profiles(element, field_type),
-                     "type": field_type,
-                     "recommended": is_recommended_field,
-                     "required": is_required_field
-                     }
-
-            if field_type == "Reference" and len(self.get_referenced_mii_profiles(element, field_type)) == 0:
-                self.logger.warning(f"Discarding field: {element['id']} - No mii profiles that match referenced profiles, parent profile is {profile['url']}")
-                continue
-
-            self.insert_field_to_tree(field_tree, field)
-
-        profile_detail["fields"] = field_tree["children"]
-
-        return profile_detail
+                self.__logger.debug(f"Profile [url={sd.get('url')}] did not match conditions => Skipping")
+        return profile_details